NutritionInformation as ValueObject
